added authorization helper, updated routes
